refactor(linkify): restructure component into multiple folders for better organization and scalability
