+++ src/components/UserTooltip/UserTooltip.tsx
@@ -7,7 +7,7 @@ import Link from "next/link";
 import { memo, PropsWithChildren, useEffect, useMemo } from "react";
 import FollowButton from "../customComponents/TrendsSidebar/components/whoToFollow/components/FollowButton";
 import UserAvatar from "../customComponents/UserAvatar";
-import Linkify from "../Linkify";
+import Linkify from "../linkify/Linkify";
 import {
   Tooltip,
   TooltipContent,
